+++ musica/mechanism_configuration/branched.py
@@ -1,9 +1,9 @@
 from typing import Optional, Any, Dict, List, Union, Tuple
 from musica import _Branched, _ReactionComponent
-from .mechanism_configuration_phase import Phase
-from .mechanism_configuration_species import Species
-from .mechanism_configuration_reactions import ReactionComponentSerializer
-from .mechanism_configuration_utils import add_other_properties, remove_empty_keys
+from .phase import Phase
+from .species import Species
+from .reactions import ReactionComponentSerializer
+from .utils import add_other_properties, remove_empty_keys
 
 
 class Branched(_Branched):
