+++ musica/mechanism_configuration/first_order_loss.py
@@ -1,0 +1,67 @@
+from typing import Optional, Any, Dict, List, Union, Tuple
+from musica import _FirstOrderLoss, _ReactionComponent
+from .phase import Phase
+from .species import Species
+from .reactions import ReactionComponentSerializer
+from .utils import add_other_properties, remove_empty_keys
+
+
+class FirstOrderLoss(_FirstOrderLoss):
+    """
+    A class representing a first-order loss reaction rate constant.
+
+    Attributes:
+        name (str): The name of the first-order loss reaction rate constant.
+        scaling_factor (float): The scaling factor for the first-order loss rate constant.
+        reactants (List[Union[Species, Tuple[float, Species]]]): A list of reactants involved in the reaction.
+        gas_phase (Phase): The gas phase in which the reaction occurs.
+        other_properties (Dict[str, Any]): A dictionary of other properties of the first-order loss reaction rate constant.
+    """
+
+    def __init__(
+        self,
+        name: Optional[str] = None,
+        scaling_factor: Optional[float] = None,
+        reactants: Optional[List[Union[Species, Tuple[float, Species]]]] = None,
+        gas_phase: Optional[Phase] = None,
+        other_properties: Optional[Dict[str, Any]] = None,
+    ):
+        """
+        Initializes the FirstOrderLoss object with the given parameters.
+
+        Args:
+            name (str): The name of the first-order loss reaction rate constant.
+            scaling_factor (float): The scaling factor for the first-order loss rate constant.
+            reactants (List[Union[Species, Tuple[float, Species]]]): A list of reactants involved in the reaction.
+            gas_phase (Phase): The gas phase in which the reaction occurs.
+            other_properties (Dict[str, Any]): A dictionary of other properties of the first-order loss reaction rate constant.
+        """
+        super().__init__()
+        self.name = name if name is not None else self.name
+        self.scaling_factor = scaling_factor if scaling_factor is not None else self.scaling_factor
+        self.reactants = (
+            [
+                (
+                    _ReactionComponent(r.name)
+                    if isinstance(r, Species)
+                    else _ReactionComponent(r[1].name, r[0])
+                )
+                for r in reactants
+            ]
+            if reactants is not None
+            else self.reactants
+        )
+        self.gas_phase = gas_phase.name if gas_phase is not None else self.gas_phase
+        self.other_properties = other_properties if other_properties is not None else self.other_properties
+
+    @staticmethod
+    def serialize(cls) -> Dict:
+        serialize_dict = {
+            "type": "FIRST_ORDER_LOSS",
+            "name": cls.name,
+            "scaling factor": cls.scaling_factor,
+            "reactants": ReactionComponentSerializer.serialize_list_reaction_components(cls.reactants),
+            "gas phase": cls.gas_phase,
+        }
+        add_other_properties(serialize_dict, cls.other_properties)
+        return remove_empty_keys(serialize_dict)
