--- musica/mechanism_configuration/mechanism_configuration_tunneling.py
@@ -1,103 +1,0 @@
-from typing import Optional, Any, Dict, List, Union, Tuple
-from musica import _Tunneling, _ReactionComponent
-from .mechanism_configuration_phase import Phase
-from .mechanism_configuration_species import Species
-from .mechanism_configuration_reactions import ReactionComponentSerializer
-from .mechanism_configuration_utils import add_other_properties, remove_empty_keys
-
-
-class Tunneling(_Tunneling):
-    """
-    A class representing a quantum tunneling reaction rate constant.
-
-    k = A * exp( -B / T ) * exp( C / T^3 )
-
-    where:
-        k = rate constant
-        A = pre-exponential factor [(mol m-3)^(n-1)s-1]
-        B = tunneling parameter [K^-1]
-        C = tunneling parameter [K^-3]
-        T = temperature [K]
-        n = number of reactants
-
-    Attributes:
-        name (str): The name of the tunneling reaction rate constant.
-        A (float): Pre-exponential factor [(mol m-3)^(n-1)s-1].
-        B (float): Tunneling parameter [K^-1].
-        C (float): Tunneling parameter [K^-3].
-        reactants (List[Union[Species, Tuple[float, Species]]]): A list of reactants involved in the reaction.
-        products (List[Union[Species, Tuple[float, Species]]]): A list of products formed in the reaction.
-        gas_phase (Phase): The gas phase in which the reaction occurs.
-        other_properties (Dict[str, Any]): A dictionary of other properties of the tunneling reaction rate constant.
-    """
-
-    def __init__(
-        self,
-        name: Optional[str] = None,
-        A: Optional[float] = None,
-        B: Optional[float] = None,
-        C: Optional[float] = None,
-        reactants: Optional[List[Union[Species, Tuple[float, Species]]]] = None,
-        products: Optional[List[Union[Species, Tuple[float, Species]]]] = None,
-        gas_phase: Optional[Phase] = None,
-        other_properties: Optional[Dict[str, Any]] = None,
-    ):
-        """
-        Initializes the Tunneling object with the given parameters.
-
-        Args:
-            name (str): The name of the tunneling reaction rate constant.
-            A (float): Pre-exponential factor [(mol m-3)^(n-1)s-1].
-            B (float): Tunneling parameter [K^-1].
-            C (float): Tunneling parameter [K^-3].
-            reactants (List[Union[Species, Tuple[float, Species]]]): A list of reactants involved in the reaction.
-            products (List[Union[Species, Tuple[float, Species]]]): A list of products formed in the reaction.
-            gas_phase (Phase): The gas phase in which the reaction occurs.
-            other_properties (Dict[str, Any]): A dictionary of other properties of the tunneling reaction rate constant.
-        """
-        super().__init__()
-        self.name = name if name is not None else self.name
-        self.A = A if A is not None else self.A
-        self.B = B if B is not None else self.B
-        self.C = C if C is not None else self.C
-        self.reactants = (
-            [
-                (
-                    _ReactionComponent(r.name)
-                    if isinstance(r, Species)
-                    else _ReactionComponent(r[1].name, r[0])
-                )
-                for r in reactants
-            ]
-            if reactants is not None
-            else self.reactants
-        )
-        self.products = (
-            [
-                (
-                    _ReactionComponent(p.name)
-                    if isinstance(p, Species)
-                    else _ReactionComponent(p[1].name, p[0])
-                )
-                for p in products
-            ]
-            if products is not None
-            else self.products
-        )
-        self.gas_phase = gas_phase.name if gas_phase is not None else self.gas_phase
-        self.other_properties = other_properties if other_properties is not None else self.other_properties
-
-    @staticmethod
-    def serialize(cls) -> Dict:
-        serialize_dict = {
-            "type": "TUNNELING",
-            "name": cls.name,
-            "A": cls.A,
-            "B": cls.B,
-            "C": cls.C,
-            "reactants": ReactionComponentSerializer.serialize_list_reaction_components(cls.reactants),
-            "products": ReactionComponentSerializer.serialize_list_reaction_components(cls.products),
-            "gas phase": cls.gas_phase,
-        }
-        add_other_properties(serialize_dict, cls.other_properties)
-        return remove_empty_keys(serialize_dict)
