+++ musica/mechanism_configuration/phase_transfer.py
@@ -1,8 +1,8 @@
 from typing import Optional, Any, Dict, List, Union, Tuple
 from musica import _SimpolPhaseTransfer, _ReactionComponent
-from .mechanism_configuration_phase import Phase
-from .mechanism_configuration_species import Species
-from .mechanism_configuration_utils import add_other_properties, remove_empty_keys
+from .phase import Phase
+from .species import Species
+from .utils import add_other_properties, remove_empty_keys
 
 
 class SimpolPhaseTransfer(_SimpolPhaseTransfer):
