+++ musica/mechanism_configuration/reactions.py
@@ -1,7 +1,7 @@
 from typing import Optional, Any, Dict, List, Union, Tuple
 from musica import (_ReactionType, _Reactions, _ReactionsIterator)
-from .mechanism_configuration_species import Species, _Species
-from .mechanism_configuration_utils import remove_empty_keys
+from .species import Species, _Species
+from .utils import remove_empty_keys
 
 
 class ReactionType(_ReactionType):
